Extract and reuse shared helpers in the shared scope
